Makin' a scene.
Taking a scene-based approach to the game. Scenes are such things as the title screen, the overworld screen, the battle screen... basically different "modes" for the game to switch between as it sees fit. Each scene will handle input and display functionality, along with just generally the overal logic of the game. This does mean, though, that we'll need to make sure that each scene transition passes vital variables (party stats, stuff like that) from the old scene to the new.