Images for everyone!
An image each has been added to the title and battle screens. The title screen image is a dummy screen, formed by mashing up a mockup made by Alzoru and some hastily and sloppily added in text. Hooray! The battle image is also a placeholder, which is why it looks so... THERE, I guess is the word?