Bare-bones combat system
I did some comment touch-up here, but most of the changes are to combat. As opposed to an inert BattleScene, now it actually functions, though all you can do is keep exchanging one-damage blows with the enemy until one of you falls over. And I have yet to program falling over. Also, the smaller background image from before, which would spend most of the time obscured by combat windows, is now a transparent background image that shows through the windows. As with the title screen, it's a dummy image.