HP regen is active
I have programmed in the HP regeneration sensation that is sweeping the nation throughout the generations. Doesn't apply to the dead, but does apply to the unconscious, so people can revive from unconsciousness this way. Not enemies, though, since they're considered dead when HP reaches zero, for the sake of convenience.