Object Tree Navigation and Test
The "object tree" I refer to in this case is the tree representing the various parts of a mod, as seen in DevScenes. As a test of this, the test battle now substitutes the first up-to-four characters found into the party. Also, comments have been revamped somewhat, most notably in DevScenes.